+++ Coimbra.Editor/GUIDrawers/SelectableLabelDrawer.cs
@@ -1,0 +1,30 @@
+﻿using UnityEditor;
+using UnityEngine;
+
+namespace Coimbra.Editor
+{
+    /// <summary>
+    /// Drawer for <see cref="SelectableLabelAttribute"/>.
+    /// </summary>
+    [CustomPropertyDrawer(typeof(SelectableLabelAttribute))]
+    public sealed class SelectableLabelDrawer : PropertyDrawer
+    {
+        /// <inheritdoc/>
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            return EditorGUIUtility.singleLineHeight;
+        }
+
+        /// <inheritdoc/>
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            using EditorGUI.PropertyScope propertyScope = new EditorGUI.PropertyScope(position, label, property);
+            position = EditorGUI.PrefixLabel(position, propertyScope.content);
+
+            using (new EditorGUI.IndentLevelScope(-EditorGUI.indentLevel))
+            {
+                EditorGUI.SelectableLabel(position, property.stringValue);
+            }
+        }
+    }
+}