+++ Coimbra.Tests/EventSystemTests.cs
@@ -113,7 +113,7 @@ namespace Coimbra.Tests
         [Test]
         public void RemoveListener_Empty()
         {
-            _eventSystem.RemoveListener(EventHandle.Create(typeof(TestEvent)));
+            _eventSystem.RemoveListener(EventHandle.Create(_eventSystem, typeof(TestEvent)));
             _eventSystem.Invoke(this, new TestEvent());
             LogAssert.NoUnexpectedReceived();
         }