+++ Coimbra/AssemblyInfo.cs
@@ -4,6 +4,7 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Coimbra.Editor.Tests")]
 [assembly: InternalsVisibleTo("Coimbra.Editor.UPM")]
 [assembly: InternalsVisibleTo("Coimbra.Services.Editor")]
+[assembly: InternalsVisibleTo("Coimbra.Services.Events.Editor")]
 [assembly: InternalsVisibleTo("Coimbra.Services.PlayerLoopEvents")]
 [assembly: InternalsVisibleTo("Coimbra.Services.Pooling")]
 [assembly: InternalsVisibleTo("Coimbra.Tests")]