+++ Coimbra/DisableResizeAttribute.cs
@@ -3,7 +3,7 @@
 namespace Coimbra
 {
     /// <summary>
-    /// Apply to a <see cref="SerializableList{T}"/> or <see cref="SerializableDictionary{TKey,TValue}"/> to lock its size in the inspector.
+    /// Apply to a <see cref="SerializableDictionary{TKey,TValue}"/> to lock its size in the inspector.
     /// </summary>
     [AttributeUsage(AttributeTargets.Field)]
     public sealed class DisableResizeAttribute : Attribute { }