+++ Coimbra/PropertyAttributes/SelectableLabelAttribute.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace Coimbra
+{
+    /// <summary>
+    /// Add this to a string field to make it a readonly selectable label in the inspector.
+    /// </summary>
+    public sealed class SelectableLabelAttribute : PropertyAttribute { }
+}